fix+test: SD-1 security lint, LL-1, CE-1, ID-1, UG-1, RP-1, BA-1, RS-1, CV-1 (pass-18 low/medium batch)
SD-1: new scripts/lint-security-doc.py parses chrome-extension/manifest.json
host_permissions and SECURITY.md, asserts every HTTPS host in the
manifest is referenced in the doc. Wired into `task test`. Prevents
SECURITY.md drift when a future commit adds an outbound fetch target.

LL-1: extend the L-9 comment in _getPrimary's idle callback to call out
that BOTH `_clearMetricIdleId = null` AND `_clearingMetric = false` MUST
precede `set_string('')`. Either re-order silently locks the
orphan-recovery path.

CE-1: load_contents callback now special-cases Gio.IOErrorEnum.CANCELLED
so the normal cancel-on-rapid-file-change path doesn't spam journalctl
with "failed to read cache". Other errors still log.

ID-1: install.sh comment documents that --delete wipes user edits under
$SERVER_DIR/chrome-extension/ — intentional, not a bug.

UG-1: install.sh comment notes the uninstall glob matches by filename
because we own the `claude-usage` IconName.

RP-1: render-panel-screenshot.py PCT_VALUE now `min(99, threshold_warning
+ 4)`. Future schema edit that raises threshold_warning above 96 no
longer renders "100%+".

BA-1: usage-server.py _buffered_at bound check uses `int(time.time() *
1000)` to match the pure-int style of _timestamp's check. Cosmetic.

RS-1: install.sh now fails loudly at start if rsync isn't on PATH,
rather than partway through the chrome-extension copy.

CV-1: new test_cache_reset_on_non_dict_prev covers the CM-1 reset path.
Before pass-17 CM-1, a corrupt usage.json containing `[]`/scalar/null
crashed every POST. Test asserts the POST now succeeds against four
non-dict shapes. Suite: 93 → 94.

+++ server/usage-server.py
@@ -243,11 +243,14 @@ def _validate(body):
     # V-1 (pass-17): _buffered_at is whitelisted in _VALID_TOP_KEYS but was
     # never bounded — asymmetric with siblings (_timestamp, _scrape_fail_count
     # all enforce plausibility). It's epoch-ms per the Chrome ext.
+    # BA-1 (pass-18): int() the comparison anchor to mirror _timestamp's
+    # pure-int comparison style (cosmetic — floats at epoch-ms magnitude
+    # have ~0.0002 ms precision, well under bound resolution).
     ba = body.get('_buffered_at')
     if ba is not None:
         if isinstance(ba, bool) or not isinstance(ba, (int, float)):
             return "'_buffered_at' must be a number"
-        now_ms = time.time() * 1000
+        now_ms = int(time.time() * 1000)
         if not (now_ms - 365 * 86400 * 1000 < ba < now_ms + 86400 * 1000):
             return "'_buffered_at' implausibly far from server time"
     return None
